+++ backend/business/internal/service/impl/stats_service.go
@@ -23,7 +23,7 @@ func NewStatsServiceImpl(statsRepo repository.StatsRepo) *StatsServiceImpl {
 
 // GetTotalPosts は投稿の総数を取得します（M3-7-1）。
 // 掲載投稿数は投稿テーブルのレコード数
-func (s *StatsServiceImpl) GetTotalPosts(ctx context.Context, businessID int64) (interface{}, error) {
+func (s *StatsServiceImpl) GetTotalPosts(ctx context.Context, businessID int32) (interface{}, error) {
 	if businessID <= 0 {
 		return nil, errors.NewAPIError(errors.ErrInvalidInput, "businessId must be greater than 0")
 	}
@@ -34,14 +34,14 @@ func (s *StatsServiceImpl) GetTotalPosts(ctx context.Context, businessID int64) 
 	}
 
 	return &domain.StatsResponse{
-		Total: count,
+		Total: int32(count),
 		Label: "Total Posts",
 	}, nil
 }
 
 // GetTotalReactions はリアクションの総数を取得します（M3-7-2）。
 // リアクション数は同一ユーザーが同一投稿に複数回リアクションできないため COUNT(DISTINCT reaction)
-func (s *StatsServiceImpl) GetTotalReactions(ctx context.Context, businessID int64) (interface{}, error) {
+func (s *StatsServiceImpl) GetTotalReactions(ctx context.Context, businessID int32) (interface{}, error) {
 	if businessID <= 0 {
 		return nil, errors.NewAPIError(errors.ErrInvalidInput, "businessId must be greater than 0")
 	}
@@ -52,14 +52,14 @@ func (s *StatsServiceImpl) GetTotalReactions(ctx context.Context, businessID int
 	}
 
 	return &domain.StatsResponse{
-		Total: count,
+		Total: int32(count),
 		Label: "Total Reactions",
 	}, nil
 }
 
 // GetTotalViews は閲覧数の総数を取得します（M3-7-3）。
 // SSOT Rules: 表示回数は各投稿のビューカウントの合計
-func (s *StatsServiceImpl) GetTotalViews(ctx context.Context, businessID int64) (interface{}, error) {
+func (s *StatsServiceImpl) GetTotalViews(ctx context.Context, businessID int32) (interface{}, error) {
 	if businessID <= 0 {
 		return nil, errors.NewAPIError(errors.ErrInvalidInput, "businessId must be greater than 0")
 	}
@@ -70,14 +70,14 @@ func (s *StatsServiceImpl) GetTotalViews(ctx context.Context, businessID int64) 
 	}
 
 	return &domain.StatsResponse{
-		Total: count,
+		Total: int32(count),
 		Label: "Total Views",
 	}, nil
 }
 
 // GetEngagementRate はエンゲージメント率を計算します（M3-7-4）。
 // SSOT Rules: エンゲージメント率 = (リアクション数 + ビュー数) / (投稿数 * 100)
-func (s *StatsServiceImpl) GetEngagementRate(ctx context.Context, businessID int64) (interface{}, error) {
+func (s *StatsServiceImpl) GetEngagementRate(ctx context.Context, businessID int32) (interface{}, error) {
 	if businessID <= 0 {
 		return nil, errors.NewAPIError(errors.ErrInvalidInput, "businessId must be greater than 0")
 	}
@@ -95,9 +95,9 @@ func (s *StatsServiceImpl) GetEngagementRate(ctx context.Context, businessID int
 	}
 
 	return &domain.EngagementResponse{
-		PostCount:      postCount,
-		ReactionCount:  reactionCount,
-		ViewCount:      viewCount,
+		PostCount:      int32(postCount),
+		ReactionCount:  int32(reactionCount),
+		ViewCount:      int32(viewCount),
 		EngagementRate: engagementRate,
 	}, nil
 }
